refactor: use nested route
