+++ openpbl-landing/src/pages/Project/CreateProject/Info/index.jsx
@@ -25,21 +25,21 @@ class EditInfo extends React.PureComponent {
     const {pid} = this.state
     return (
       <DocumentTitle title="Project">
-        <div style={{backgroundColor: '#f2f4f5', minHeight: '100vh',}}>
-          <PageHeader
-            className="site-page-header"
-            onBack={() => this.back()}
-            title="返回"
-            subTitle="我的项目"
-          />
           <div
             style={{
-              paddingTop: '20px',
               paddingLeft: '20px',
               paddingRight: '20px',
-              paddingBottom: '20px'
+              paddingBottom: '20px',
+              maxWidth: '1200px',
+              margin: 'auto',
             }}
           >
+            <PageHeader
+              className="site-page-header"
+              onBack={() => this.back()}
+              title="返回"
+              subTitle="我的项目"
+            />
             <Card className="edit-card">
               <Steps current={0} style={{marginLeft: '20px'}}>
                 <Step title="填写信息"/>
@@ -49,7 +49,6 @@ class EditInfo extends React.PureComponent {
               <InfoEditPage pid={pid}/>
             </Card>
           </div>
-        </div>
       </DocumentTitle>
     )
   }
