+++ openpbl-landing/src/pages/Project/CreateProject/Outline/index.less
@@ -1,5 +1,3 @@
 .edit-card {
-  max-width: 1200px;
-  margin: auto;
   text-align: left;
 }
