+++ openpbl-landing/src/pages/Project/MyProject/index.jsx
@@ -1,32 +1,119 @@
 import React from 'react';
 import DocumentTitle from 'react-document-title';
 
-import GlobalHeader from '../../component/GlobalHeader/GlobalHeader';
-import MenuBar from "../component/MenuBar";
 import localStorage from "localStorage";
+import {Switch, Route, Link, Redirect} from 'react-router-dom'
+import PublishedProject from "../PublishedProject";
+import EditingProject from "../EditingProject";
+import FinishedProject from "../FinishedProject";
+import LearningProject from "../LearningProject";
+import {Affix, Button, Layout, Menu} from "antd";
+import {CheckCircleOutlined, CheckOutlined, CopyOutlined, HighlightOutlined, SyncOutlined} from "@ant-design/icons";
+import ProjectApi from "../../../api/ProjectApi";
 
 class MyProject extends React.PureComponent {
   state = {
     type: localStorage.getItem('type'),
   }
+  createProject = e => {
+    ProjectApi.createProject()
+      .then((res)=>{
+        if (res.data.code === 200) {
+          window.open(`/project/${res.data.data}/info/edit`)
+        }
+      })
+      .catch((e)=>{console.log(e)})
+  }
   render() {
     const {type} = this.state
     return (
-      <DocumentTitle title="Project">
-        <div style={{minHeight: '100vh'}}>
-          <GlobalHeader current="my-project"/>
-          <div>
-            {type === 'teacher' ?
-              <MenuBar menu="editing"/>
-              :
-              <MenuBar menu="learning"/>
-            }
-          </div>
-        </div>
+      <DocumentTitle title="My Project">
+          <Layout>
+            <Affix offsetTop={0}>
+              <Layout.Sider
+                breakpoint="lg"
+                collapsedWidth="0"
+                theme="light"
+                style={{backgroundColor: '#f2f4f5'}}
+              >
+                <Menu
+                  defaultSelectedKeys={['published']}
+                  className="menu-bar"
+                  mode="inline"
+                >
+                  {type === 'teacher' ?
+                    <>
+                      <Menu.Item key="published" icon={<CheckCircleOutlined/>}>
+                        <Link to="/my-project/published">
+                          已发布
+                        </Link>
+                      </Menu.Item>
+                      <Menu.Item key="editing" icon={<HighlightOutlined/>}>
+                        <Link to="/my-project/editing">
+                          未发布
+                        </Link>
+                      </Menu.Item>
+                      <Menu.Item key="finished" icon={<CopyOutlined/>}>
+                        <Link to="/my-project/finished">
+                          已结束
+                        </Link>
+                      </Menu.Item>
+                    </>
+                    :
+                    <>
+                      <Menu.Item key="learning" icon={<SyncOutlined/>}>
+                        <Link to="/my-project/learning">
+                          进行中
+                        </Link>
+                      </Menu.Item>
+                      <Menu.Item key="finished" icon={<CheckOutlined/>}>
+                        <Link to="/my-project/finished">
+                          已完成
+                        </Link>
+                      </Menu.Item>
+                    </>
+                  }
+                </Menu>
+                {type === 'teacher' ?
+                  <Button
+                    type='primary'
+                    shape='round'
+                    size="large"
+                    onClick={this.createProject}
+                    style={{
+                      width: '180px',
+                      margin: '10px',
+                    }}
+                  >创建项目</Button>
+                  :
+                  null
+                }
+              </Layout.Sider>
+            </Affix>
+            <Layout.Content style={{marginLeft: '10px', marginTop: '5px'}}>
+              {type === 'teacher' ?
+                <Switch>
+                  <Route exact path="/my-project" render={()=>(
+                    <Redirect to="/my-project/published" />
+                  )}/>
+                  <Route exact path="/my-project/published" component={PublishedProject}/>
+                  <Route exact path="/my-project/editing" component={EditingProject}/>
+                  <Route exact path="/my-project/finished" component={FinishedProject}/>
+                </Switch>
+                :
+                <Switch>
+                  <Route exact path="/my-project" render={()=>(
+                    <Redirect to="/my-project/learning" />
+                  )}/>
+                  <Route exact path="/my-project/learning" component={LearningProject}/>
+                  <Route exact path="/my-project/finished" component={FinishedProject}/>
+                </Switch>
+              }
+            </Layout.Content>
+          </Layout>
       </DocumentTitle>
     );
   }
 }
 
-
-export default MyProject;
+  export default MyProject;
