+++ openpbl-landing/src/pages/Project/PublicProject/index.jsx
@@ -5,7 +5,7 @@ import CarouselPBL from './component/CarouselPBL';
 class PublicProject extends React.PureComponent {
   render() {
     return (
-      <div style={{ minHeight: '100vh',margin: 'auto', maxWidth: '1200px' }}>
+      <div style={{maxWidth: '1200px', margin: 'auto'}}>
         <CarouselPBL/><br />
         <div >
           <ProjectList mode="public"/>
