+++ openpbl-landing/src/pages/User/Login/index.jsx
@@ -6,7 +6,6 @@ import localStorage from 'localStorage';
 import QueueAnim from 'rc-queue-anim';
 
 import './login.less';
-import GlobalHeader from "../../component/GlobalHeader/GlobalHeader";
 
 
 const {TabPane} = Tabs;
@@ -39,7 +38,6 @@ class Login extends React.PureComponent {
     return (
       <DocumentTitle title="Login">
         <div style={{height: '100vh'}}>
-          <GlobalHeader current="login"/>
           <QueueAnim delay={100}>
             <Card className="login-card" key="1">
               <Tabs defaultActiveKey={type} onChange={this.changeTab}>
